Restructred the cli to be modular and added solana-program
